+++ NEWS.md
@@ -94,6 +94,7 @@ the new router layout. Items must be wrapped in a `shiny::tagList()`.
 - `updateF7App` can now also handle changes in app theme (ios or md), dark mode, and color.
 - `f7Fabs()` has a new argument `global` that can be used to make FABs persistent across different tabs in `f7TabLayout()`. 
 - `f7ExpandableCard()` has a new argument `buttonColor` that can be used to control the color of the close button.
+- `f7Login()` has a new argument `module` that can, optionally, be set to `FALSE` for more flexibility. For example, this allows you to use `f7Login()` inside your own modules, or without the provided `f7LoginServer()` module.
 - Fix various issues in documentation.
 - Include new vignettes.